// frontend/src/app/gyms/+state/enums/gym-status.enum.ts
export enum GymStatusEnum {
  pending = 'pending',
  loading = 'loading',
  loadSuccess = 'loadSuccess',
  loadError = 'loadError',
  loadDetailsSuccess = 'loadDetailsSuccess',
  loadDetailsFailure = 'loadDetailsFailure',
  createSuccess = 'createSuccess',
  createFailure = 'createFailure',
  updateSuccess = 'updateSuccess',
  updateFailure = 'updateFailure',
  deleteSuccess = 'deleteSuccess',
  deleteFailure = 'deleteFailure',
  approveSuccess = 'approveSuccess',
  approveFailure = 'approveFailure',
}